Refactor some dependencies
