+++ src/main.rs
@@ -4,20 +4,9 @@ mod models;
 mod packet_handlers;
 mod server;
 
-use game_state::patchwork::{NewMapMessage, PatchworkStateOperations};
+use services::game_state::patchwork::{NewMapMessage, PatchworkStateOperations};
 
-use packet_handlers::packet_router;
-
-use models::map;
-use models::minecraft_types;
-use models::packet;
-use models::translation;
-
-use services::game_state;
 use services::instance::ServiceInstance;
-use services::keep_alive;
-use services::messenger;
-use services::packet_processor;
 
 use std::env;
 use std::thread;
@@ -44,32 +33,32 @@ fn main() {
 
     define_services!(
         (
-            module: game_state::player::start,
+            module: services::game_state::player::start,
             name: player_state,
             dependencies: [messenger]
         ),
         (
-            module: game_state::block::start,
+            module: services::game_state::block::start,
             name: block_state,
             dependencies: [messenger]
         ),
         (
-            module: game_state::patchwork::start,
+            module: services::game_state::patchwork::start,
             name: patchwork_state,
             dependencies: [messenger, inbound_packet_processor, player_state]
         ),
         (
-            module: messenger::start,
+            module: services::messenger::start,
             name: messenger,
             dependencies: []
         ),
         (
-            module: packet_processor::start_inbound,
+            module: services::packet_processor::start_inbound,
             name: inbound_packet_processor,
             dependencies: [messenger, player_state, block_state, patchwork_state]
         ),
         (
-            module: keep_alive::start,
+            module: services::keep_alive::start,
             name: keep_alive,
             dependencies: [messenger]
         )
@@ -84,7 +73,7 @@ fn main() {
     patchwork_state
         .sender()
         .send(PatchworkStateOperations::New(NewMapMessage {
-            peer: map::Peer {
+            peer: models::map::Peer {
                 port: peer_port,
                 address: peer_address,
             },
