+++ src/models.rs
@@ -1,12 +1,13 @@
 #[macro_use]
 mod packet_macros;
 pub mod map;
-mod minecraft_protocol;
+pub mod minecraft_protocol;
 pub mod minecraft_types;
 pub mod packet;
 pub mod translation;
 
-use super::game_state;
-use super::messenger;
-use super::packet_processor;
+use super::services::game_state;
+use super::services::messenger;
+use super::services::packet_processor;
+
 use super::server;
