+++ src/models/minecraft_protocol.rs
@@ -1,9 +1,9 @@
 extern crate byteorder;
 
-use super::minecraft_types::{read_var_int, ChunkSection};
+use super::minecraft_types::ChunkSection;
 use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
 use std::cmp::{max, min};
-use std::io::{Read, Write};
+use std::io::{Error, Read, Write};
 
 const PALETTE_SIZE: i64 = 14; // We don't define our own palette, so we just use the default all blocks palette which is 14 bits
 
@@ -11,6 +11,7 @@ pub trait MinecraftProtocolReader {
     fn read_unsigned_short(&mut self) -> u16;
     fn read_short(&mut self) -> i16;
     fn read_var_int(&mut self) -> i32;
+    fn try_read_var_int(&mut self) -> Result<i32, Error>;
     fn read_long(&mut self) -> i64;
     fn read_string(&mut self) -> String;
     fn read_u_128(&mut self) -> u128;
@@ -43,98 +44,17 @@ pub trait MinecraftProtocolWriter {
     fn write_boolean(&mut self, v: bool);
 }
 
-//We now have a functional though messy implementation of writing any block
-pub fn write_chunk_section<S: Write>(stream: &mut S, v: ChunkSection) {
-    stream.write_u_byte(v.bits_per_block);
-    stream.write_var_int(v.data_array_length);
-    let mut long: i64 = 0;
-    for i in 0..4096 {
-        let block_to_place = v.block_ids[i as usize] as i64;
-        let offset = (PALETTE_SIZE * i) % 64;
-        long += block_to_place << offset;
-        if ((i * PALETTE_SIZE) % 64) >= 64 - PALETTE_SIZE {
-            stream.write_long(long);
-            long = block_to_place >> (64 - offset);
-        }
-    }
-    for _ in 0..2048 {
-        stream
-            .write_u8(!0b0)
-            .expect("could not write max block light"); //write max block light
-    }
-    for _ in 0..2048 {
-        stream
-            .write_u8(!0b0)
-            .expect("could not write max sky light"); //write max sky light
-    }
-}
-
-pub fn read_chunk_section<S: Read>(stream: &mut S) -> ChunkSection {
-    let bits_per_block = stream.read_u_byte();
-    if bits_per_block != PALETTE_SIZE as u8 {
-        panic!("Cannot read palettes");
-    }
-    let data_array_length = stream.read_var_int();
-    if data_array_length != 896 {
-        panic!("Got unexpected data array length");
-    }
-    let mut block_ids = Vec::<i32>::new();
-    let mut long = stream.read_u64::<BigEndian>().unwrap();
-    let mut index = 0;
-    for i in 0..4096 {
-        let bits_to_read = min(64 - (index % 64), 14);
-        let left_shift = max(64 - (bits_to_read + (index % 64)), 0);
-        let right_shift = left_shift + (index % 64);
-        let mut block_id = (long << left_shift) >> right_shift;
-        if left_shift == 0 && i != 4095 {
-            long = stream.read_u64::<BigEndian>().unwrap();
-        }
-        if bits_to_read < 14 {
-            let remainder_to_read = 14 - bits_to_read;
-            let remainder = long << (64 - remainder_to_read) >> (64 - remainder_to_read);
-            block_id += remainder >> remainder_to_read;
-        };
-        block_ids.push(block_id as i32);
-        index += 14;
-    }
-    //Still ignoring these values for now
-    for _ in 0..2048 {
-        stream.read_u8().unwrap();
-    }
-    for _ in 0..2048 {
-        stream.read_u8().unwrap();
-    }
-    ChunkSection {
-        bits_per_block,
-        data_array_length,
-        block_ids,
-        block_light: Vec::<u64>::new(),
-        sky_light: Vec::<u64>::new(),
-    }
-}
-
-pub fn write_var_int<S: Write>(stream: &mut S, v: i32) {
-    let mut value = v;
-    loop {
-        let mut temp = value & 0b0111_1111;
-        value >>= 7;
-        if value != 0 {
-            temp |= 0b1000_0000;
-        }
-        stream.write_u8(temp as u8).unwrap();
-        if value == 0 {
-            break;
-        }
-    }
-}
-
 impl<T: Read> MinecraftProtocolReader for T {
     fn read_long(&mut self) -> i64 {
         self.read_i64::<BigEndian>().unwrap()
     }
 
     fn read_var_int(&mut self) -> i32 {
-        read_var_int(self).unwrap()
+        self.try_read_var_int().unwrap()
+    }
+
+    fn try_read_var_int(&mut self) -> Result<i32, Error> {
+        read_var_int(self)
     }
 
     fn read_unsigned_short(&mut self) -> u16 {
@@ -275,5 +195,109 @@ impl<T: Write> MinecraftProtocolWriter for T {
         } else {
             self.write_u8(0).unwrap()
         }
+    }
+}
+
+fn read_var_int<S: Read>(stream: &mut S) -> Result<i32, Error> {
+    let mut num_read = 0;
+    let mut result: i32 = 0;
+
+    loop {
+        let value = i32::from(stream.read_u8()?);
+        result |= (value & 0b0111_1111) << (7 * num_read);
+        num_read += 1;
+        if num_read > 5 {
+            panic!("VarInt is too big");
+        }
+        if (value & 0b1000_0000) == 0 {
+            break;
+        }
+    }
+
+    Ok(result)
+}
+
+fn write_var_int<S: Write>(stream: &mut S, v: i32) {
+    let mut value = v;
+    loop {
+        let mut temp = value & 0b0111_1111;
+        value >>= 7;
+        if value != 0 {
+            temp |= 0b1000_0000;
+        }
+        stream.write_u8(temp as u8).unwrap();
+        if value == 0 {
+            break;
+        }
+    }
+}
+
+//We now have a functional though messy implementation of writing any block
+fn write_chunk_section<S: Write>(stream: &mut S, v: ChunkSection) {
+    stream.write_u_byte(v.bits_per_block);
+    stream.write_var_int(v.data_array_length);
+    let mut long: i64 = 0;
+    for i in 0..4096 {
+        let block_to_place = v.block_ids[i as usize] as i64;
+        let offset = (PALETTE_SIZE * i) % 64;
+        long += block_to_place << offset;
+        if ((i * PALETTE_SIZE) % 64) >= 64 - PALETTE_SIZE {
+            stream.write_long(long);
+            long = block_to_place >> (64 - offset);
+        }
+    }
+    for _ in 0..2048 {
+        stream
+            .write_u8(!0b0)
+            .expect("could not write max block light"); //write max block light
+    }
+    for _ in 0..2048 {
+        stream
+            .write_u8(!0b0)
+            .expect("could not write max sky light"); //write max sky light
+    }
+}
+
+fn read_chunk_section<S: Read>(stream: &mut S) -> ChunkSection {
+    let bits_per_block = stream.read_u_byte();
+    if bits_per_block != PALETTE_SIZE as u8 {
+        panic!("Cannot read palettes");
+    }
+    let data_array_length = stream.read_var_int();
+    if data_array_length != 896 {
+        panic!("Got unexpected data array length");
+    }
+    let mut block_ids = Vec::<i32>::new();
+    let mut long = stream.read_u64::<BigEndian>().unwrap();
+    let mut index = 0;
+    for i in 0..4096 {
+        let bits_to_read = min(64 - (index % 64), 14);
+        let left_shift = max(64 - (bits_to_read + (index % 64)), 0);
+        let right_shift = left_shift + (index % 64);
+        let mut block_id = (long << left_shift) >> right_shift;
+        if left_shift == 0 && i != 4095 {
+            long = stream.read_u64::<BigEndian>().unwrap();
+        }
+        if bits_to_read < 14 {
+            let remainder_to_read = 14 - bits_to_read;
+            let remainder = long << (64 - remainder_to_read) >> (64 - remainder_to_read);
+            block_id += remainder >> remainder_to_read;
+        };
+        block_ids.push(block_id as i32);
+        index += 14;
+    }
+    //Still ignoring these values for now
+    for _ in 0..2048 {
+        stream.read_u8().unwrap();
+    }
+    for _ in 0..2048 {
+        stream.read_u8().unwrap();
+    }
+    ChunkSection {
+        bits_per_block,
+        data_array_length,
+        block_ids,
+        block_light: Vec::<u64>::new(),
+        sky_light: Vec::<u64>::new(),
     }
 }
