+++ src/models/packet.rs
@@ -1,7 +1,7 @@
 #![allow(unused_variables)]
 //The macro is much cleaner if we allow for unused variables
 use super::game_state::patchwork::{CHUNK_SIZE, ENTITY_ID_BLOCK_SIZE};
-use super::minecraft_protocol::{write_var_int, MinecraftProtocolReader, MinecraftProtocolWriter};
+use super::minecraft_protocol::{MinecraftProtocolReader, MinecraftProtocolWriter};
 use super::minecraft_types::ChunkSection;
 use super::translation::TranslationInfo;
 use std::io::{Cursor, Read, Write};
