+++ src/packet_handlers/initiation_protocols.rs
@@ -13,4 +13,4 @@ pub mod login;
 use super::game_state;
 use super::messenger;
 use super::packet;
-use super::TranslationUpdates;
+use super::translation;
