+++ src/services.rs
@@ -11,10 +11,13 @@ pub mod game_state;
 pub mod keep_alive;
 pub mod packet_processor;
 
-use super::map;
-use super::minecraft_types;
-use super::packet;
+use super::models::map;
+use super::models::minecraft_types;
+use super::models::packet;
+use super::models::translation;
+
 use super::packet_handlers;
-use super::packet_router;
+
+use super::packet_handlers::packet_router;
+
 use super::server;
-use super::translation;
